# src/main/resources/openapi/order-platform-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Spring Modulith Order Platform API",
    "description": "Contract-first API for a modular monolith order management platform.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/payments/authorize": {
      "post": {
        "tags": ["Payments"],
        "operationId": "authorizePayment",
        "summary": "Authorize payment for an order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PaymentAuthorizeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment authorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaymentResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Problem" },
          "404": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/payments/{paymentId}": {
      "get": {
        "tags": ["Payments"],
        "operationId": "getPayment",
        "summary": "Get a payment by id",
        "parameters": [
          { "$ref": "#/components/parameters/PaymentId" }
        ],
        "responses": {
          "200": {
            "description": "Payment",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaymentResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "PaymentId": {
        "name": "paymentId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "responses": {
      "Problem": {
        "description": "RFC7807 problem detail",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetail" }
          }
        }
      }
    },
    "schemas": {
...
      "PaymentAuthorizeRequest": {
        "type": "object",
        "required": ["orderId", "amount"],
        "properties": {
          "orderId": { "type": "string", "format": "uuid" },
          "amount": { "$ref": "#/components/schemas/Money" }
        }
      },
      "PaymentResponse": {
        "type": "object",
        "required": ["id", "orderId", "status", "amount", "createdAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "orderId": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["PENDING", "AUTHORIZED", "DECLINED"] },
          "amount": { "$ref": "#/components/schemas/Money" },
          "createdAt": { "type": "string", "format": "date-time" },
          "authorizedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
      "Money": {
        "type": "object",
        "required": ["amount", "currency"],
        "properties": {
          "amount": { "type": "number", "minimum": 0, "multipleOf": 0.01 },
          "currency": { "type": "string", "minLength": 3, "maxLength": 3, "example": "EUR" }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string" },
          "status": { "type": "integer", "format": "int32" },
          "detail": { "type": "string" },
          "instance": { "type": "string", "format": "uri" }
        }
      }
    }
  }
}